+++ R/repl.R
@@ -1000,4 +1000,3 @@ run_repl_loop <- function(ctx) {
 
     invisible(NULL)
 }
-